# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AP Calculus AB Tutor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Thread {
  threadId: string;
}

export interface UploadedFile {
  fileId: string;
  filename: string;
}

export interface MessageInput {
  content: string;
  fileIds?: string[];
}

export interface AssistantInfo {
  /** @nullable */
  assistantId?: string | null;
  name: string;
  status: string;
}