# lib/api-zod/src/generated/types/assistantInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AP Calculus AB Tutor API
 * OpenAPI spec version: 0.1.0
 */

export interface AssistantInfo {
  /** @nullable */
  assistantId?: string | null;
  name: string;
  status: string;
}