# lib/api-zod/src/generated/types/uploadedFile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AP Calculus AB Tutor API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadedFile {
  fileId: string;
  filename: string;
}